+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Disabled clangd inlay hints (``.clangd``)
 - Added build information (``print_build_info()``, flags: ``--build-info``, ``-u``)
 - Added a function to determine the compiler name and version (``src/platform/compiler_version.h``, ``src/platform/compiler_version.cpp``)
+- Changed file structure
 
 
 2.0.1 - 2022-11-22
