+++ src/gui/image_representation.h
@@ -30,37 +30,60 @@
  */
 /**
  * @file
- * @brief Class for drawing 2D image via buffer and OpenGL
+ * @brief Image representation
  * @author Saveliy Pototskiy (SavaLione)
- * @date 26 Nov 2022
+ * @date 22 Nov 2022
  */
-#ifndef CORE_GL_IMAGE_H
-#define CORE_GL_IMAGE_H
+#ifndef GUI_IMAGE_REPRESENTATION_H
+#define GUI_IMAGE_REPRESENTATION_H
 
-#include "core/sdl_wrapper.h"
+#include <cstdint>
+#include <cstdlib>
+#include <vector>
 
-#include "core/image_representation.h"
-#include "core/buffer_representation.h"
-
-#include <array>
-
-class gl_image : sdl_wrapper
+class image_representation
 {
 public:
-    gl_image();
-    ~gl_image();
-    void run();
+    struct vec3
+    {
+        std::int8_t r;
+        std::int8_t g;
+        std::int8_t b;
+    };
 
-protected:
-    image_representation ir;
+    struct vec4
+    {
+        std::int8_t r;
+        std::int8_t g;
+        std::int8_t b;
+        std::int8_t a;
+    };
+
+    image_representation() : _width(0), _height(0), _depth(0)
+    {
+        _image.clear();
+    }
+
+    image_representation(std::size_t const &width, std::size_t const &height, std::size_t const &depth) : _width(width), _height(height), _depth(depth)
+    {
+        std::size_t size = width * height * depth;
+        _image.resize(size, 0);
+    }
+
+    std::size_t const &width() const;
+    std::size_t const &height() const;
+    std::size_t const &depth() const;
+    std::size_t size();
+    uint8_t *data();
+    uint8_t const *const_data() const;
+    void fill_zeros();
+    void set_pixel(std::size_t x, std::size_t y, vec4 color);
 
 private:
-    /* OpenGL */
-    GLuint _vao = 0;
-    GLuint _vbo = 0;
-    GLuint _ebo = 0;
-
-    std::array<GLuint, 1> textures;
+    std::size_t _width;
+    std::size_t _height;
+    std::size_t _depth;
+    std::vector<uint8_t> _image;
 };
 
-#endif // CORE_GL_IMAGE_H
+#endif // CORE_IMAGE_REPRESENTATION_H
